+++ src/main/kotlin/com/codescene/jetbrains/components/codehealth/monitor/CodeHealthMonitorPanel.kt
@@ -10,7 +10,7 @@ import com.codescene.jetbrains.notifier.CodeHealthDetailsRefreshNotifier
 import com.codescene.jetbrains.services.GitService
 import com.codescene.jetbrains.services.cache.DeltaCacheQuery
 import com.codescene.jetbrains.services.cache.DeltaCacheService
-import com.codescene.jetbrains.services.telemetry.TelemetryService
+import com.codescene.jetbrains.services.api.telemetry.TelemetryService
 import com.codescene.jetbrains.util.Constants.CODESCENE
 import com.codescene.jetbrains.util.Log
 import com.codescene.jetbrains.util.TelemetryEvents
@@ -144,7 +144,7 @@ class CodeHealthMonitorPanel(private val project: Project) {
                 return
             }
 
-        val headCommit = GitService.getInstance(project).getHeadCommit(file)
+        val headCommit = GitService.getInstance(project).getBranchCreationCommitCode(file)
 
         val cachedDelta = DeltaCacheService.getInstance(project)
             .get(DeltaCacheQuery(path, headCommit, code))