+++ src/main/kotlin/com/codescene/jetbrains/components/settings/tab/SettingsTab.kt
@@ -2,45 +2,46 @@ package com.codescene.jetbrains.components.settings.tab
 
 import com.codescene.jetbrains.UiLabelsBundle
 import com.codescene.jetbrains.config.global.CodeSceneGlobalSettingsStore
+import com.codescene.jetbrains.services.api.AceService
 import com.intellij.openapi.options.BoundConfigurable
 import com.intellij.openapi.ui.DialogPanel
 import com.intellij.ui.dsl.builder.bindSelected
 import com.intellij.ui.dsl.builder.panel
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 
 @Suppress("DialogTitleCapitalization")
 class SettingsTab : BoundConfigurable(UiLabelsBundle.message("settingsTitle")) {
     private val settings = CodeSceneGlobalSettingsStore.getInstance().state
+    private val scope = CoroutineScope(Dispatchers.IO)
 
-    override fun createPanel(): DialogPanel = panel {
-        row {
-            checkBox(UiLabelsBundle.message("enableCodeLenses"))
-                .bindSelected(settings::enableCodeLenses)
-                .comment(UiLabelsBundle.message("enableCodeLensesComment"))
-        }
+    override fun createPanel(): DialogPanel {
+        return panel {
+            row {
+                checkBox(UiLabelsBundle.message("enableCodeLenses"))
+                    .bindSelected(settings::enableCodeLenses)
+                    .comment(UiLabelsBundle.message("enableCodeLensesComment"))
+            }
 
-        /*
-        TODO: Uncomment this code when ACE is integrated in plugin:
             row {
                 checkBox(UiLabelsBundle.message("enableAutoRefactor"))
                     .bindSelected(settings::enableAutoRefactor)
                     .comment(UiLabelsBundle.message("enableAutoRefactorComment"))
             }
-         */
 
-        /*
-        TODO: Uncomment this code when Code Health gate is integrated in plugin:
+            //TODO: remove
             row {
-                checkBox(UiLabelsBundle.message("previewCodeHealthGate"))
-                    .bindSelected(settings::previewCodeHealthGate)
-                    .comment(UiLabelsBundle.message("previewCodeHealthGateComment"))
+                checkBox("Ace Acknowledged")
+                    .bindSelected(settings::aceAcknowledged)
+                    .comment("Test, remove")
             }
-        */
 
-        row {
-            checkBox(UiLabelsBundle.message("gitignore"))
-                .bindSelected(settings::excludeGitignoreFiles)
-                .comment(UiLabelsBundle.message("gitignoreComment"))
-        }
+            row {
+                checkBox(UiLabelsBundle.message("gitignore"))
+                    .bindSelected(settings::excludeGitignoreFiles)
+                    .comment(UiLabelsBundle.message("gitignoreComment"))
+            }
 
 //        panel {
 //            groupRowsRange(UiLabelsBundle.message("server")) {
@@ -54,5 +55,13 @@ class SettingsTab : BoundConfigurable(UiLabelsBundle.message("settingsTitle")) {
 //                }
 //            }
 //        }
+        }
+    }
+
+    override fun apply() {
+        super.apply()
+        scope.launch {
+            AceService.getInstance().runPreflight(true)
+        }
     }
 }