+++ src/main/kotlin/com/codescene/jetbrains/services/htmlviewer/codehealth/PreviewThemeStyles.kt
@@ -1,4 +1,4 @@
-package com.codescene.jetbrains.codeInsight.codehealth
+package com.codescene.jetbrains.services.htmlviewer.codehealth
 
 import com.codescene.jetbrains.util.webRgba
 import com.intellij.openapi.editor.colors.EditorColorsManager
@@ -20,23 +20,23 @@ object PreviewThemeStyles {
 
         // language=CSS
         return """
-            body {
-              background-color: $backgroundColorGlobalString;
-              color: $foregroundColorGlobalString;
-              font-size: ${fontSize + 1}px;
-            }
-            a {
-              color: $linkForegroundColorString;
-            }
-            pre, code, hr {
-              background-color: $codeBackgroundColorString;
-            }
-            #function-location:hover {
-              background-color: $codeBackgroundColorString;
-            }
-            #line-number {
-              color: $disabledColorString;
-            }
-            """.trimIndent()
+            |body {
+            |  background-color: $backgroundColorGlobalString;
+            |  color: $foregroundColorGlobalString;
+            |  font-size: ${fontSize + 1}px;
+            |}
+            |a {
+            |  color: $linkForegroundColorString;
+            |}
+            |pre, code, hr {
+            |  background-color: $codeBackgroundColorString;
+            |}
+            |#function-location:hover {
+            |  background-color: $codeBackgroundColorString;
+            |}
+            |#line-number {
+            |  color: $disabledColorString;
+            |}
+        """.trimMargin().trim()
     }
 }